backend/requirements.txt
gunicorn
Flask
numpy==1.14.5
setuptools==39.1.0
tensorflow==1.10.0
h5py==2.9.0
Pillow==6.0.0
gevent
